+++ CHANGELOG.md
@@ -7,7 +7,9 @@ from version 1.20.0 onwards.
 
 ## [Unreleased]
 
-- n/a
+### Fixed
+- `more_executors.futures.*` can now be referenced after an `import more_executors`;
+  there is no need to explicitly `import more_executors.futures
 
 ## [2.2.0] - 2019-08-31